Removed debug comments.Fixed bool in reduce()
The true/false debug comments mixed up saying it was not reducing when it did and vice versa

I set false to true in the reduce function because it did reduce until it hit 1 for both. i think the mistake was I wanted it to be the same number on the numerator and denominator so it would not waste time reducing if they were the same. I might reintroduce commented out code.