Added new IF to algorithm in reduce(). Pre-release
Added a condition that when the numerator was greater than the denominator it would stop
reducing. If this were to happen it would continue the same until both are not divisible.

As the program is completely finished. This stage now takes care of comments, testing syntax, and checking for heavy redundancies.